core: create dedicated file for eigen allocation
